extract campaign listing from campaign overview
